+++ packages/transaction-pay-controller/src/index.ts
@@ -4,6 +4,7 @@ export type {
   TransactionData,
   TransactionFiatPayment,
   TransactionFiatPaymentCallback,
+  GetPaymentOverrideDataCallback,
   TransactionPayControllerActions,
   TransactionPayControllerEvents,
   TransactionPayControllerGetStateAction,